refactor: move columns to own namespace
issue: #100
